# swagger.yaml
openapi: 3.0.0
info:
  title: Mood Tracker API
  description: API for tracking mood and mental health metrics with dynamic custom scales
  version: 1.0.0
  contact:
    name: API Support
    email: support@moodtracker.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.moodtracker.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
    UserRegistration:
      type: object
      required:
        - email
        - username
        - password
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 3
          maxLength: 30
        password:
          type: string
          minLength: 8
          format: password
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
...
    ScaleWeight:
      type: object
      required:
        - scaleId
        - weight
      properties:
        scaleId:
          type: string
          format: uuid
        scaleName:
          type: string
          readOnly: true
        weight:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        isInverted:
          type: boolean
          default: false
    StabilityFormula:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
        formula:
          type: string
          description: "Formula string representation"
        isDefault:
          type: boolean
        isActive:
          type: boolean
        scaleWeights:
          type: array
          items:
            $ref: '#/components/schemas/ScaleWeight'
    CreateStabilityFormula:
      type: object
      required:
        - description
        - scaleWeights
      properties:
        description:
          type: string
          maxLength: 500
        isActive:
          type: boolean
          default: true
        scaleWeights:
          type: array
          items:
            $ref: '#/components/schemas/ScaleWeight'
    UpdateStabilityFormula:
      type: object
      properties:
        description:
          type: string
          maxLength: 500
        isActive:
          type: boolean
        scaleWeights:
          type: array
          items:
            $ref: '#/components/schemas/ScaleWeight'
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
        errors:
          type: array
          items:
            type: string
security:
  - bearerAuth: []
paths:
  /api/users/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/login:
    post:
      tags:
        - Authentication
      summary: Login a user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/stability-formulas:
    get:
      tags:
        - Stability Formulas
      summary: Get all stability formulas
      responses:
        '200':
          description: List of stability formulas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StabilityFormula'
        '401':
          description: Unauthorized
    post:
      tags:
        - Stability Formulas
      summary: Create a new stability formula
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStabilityFormula'
      responses:
        '201':
          description: Stability formula created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StabilityFormula'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
  /api/stability-formulas/active:
    get:
      tags:
        - Stability Formulas
      summary: Get active stability formula
      responses:
        '200':
          description: Active stability formula
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StabilityFormula'
        '401':
          description: Unauthorized
        '404':
          description: No active formula found
  /api/stability-formulas/{id}:
    get:
      tags:
        - Stability Formulas
      summary: Get a stability formula by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Stability formula found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StabilityFormula'
        '401':
          description: Unauthorized
        '404':
          description: Stability formula not found
    put:
      tags:
        - Stability Formulas
      summary: Update a stability formula
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStabilityFormula'
      responses:
        '200':
          description: Stability formula updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StabilityFormula'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Stability formula not found
    delete:
      tags:
        - Stability Formulas
      summary: Delete a stability formula
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Stability formula deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Stability formula not found
...